Remove unnecessary type cast
Commit UUID: a43b8069-9ef4-4e1f-a916-786b0a1d54ad

+++ trunk/simplefs_dbg.c
@@ -240,7 +240,7 @@ get_key_ino(struct db_key *k)
     if (arg == NULL)
         return 2;
 
-    k->ino = (uint64_t)strtoul(arg, NULL, 10);
+    k->ino = strtoul(arg, NULL, 10);
 
     free(arg);
